requirements.txt
azure-ai-documentintelligence==1.0.0b1
openai==0.28.1